Fix synchronisation
